+++ src/components/tabbar/tabbar.js
@@ -32,7 +32,7 @@ type TabbarProps = {
 };
 
 export class Tabbar extends React.Component<TabbarProps, TabbarState> {
-    state: TabbarState = {
+    state = {
         selectedItem: 0,
     };
     render() {
@@ -41,13 +41,14 @@ export class Tabbar extends React.Component<TabbarProps, TabbarState> {
             <View style={styles.tabbar}>
                 {items.map((item, index) => (
                     <TabbarItem
+                        key={`tabbar-item-${index}`}
                         itemState={
                             index === this.state.selectedItem
                                 ? "active"
                                 : "inactive"
                         }
                         itemType={item}
-                        onPress={() => {
+                        onClick={() => {
                             this.setState({selectedItem: index});
                             onSelect(item);
                         }}
